fix includes and guard names
