+++ algorithms/src/Kokkos_NestedSort.hpp
@@ -21,7 +21,7 @@
 #define KOKKOS_IMPL_PUBLIC_INCLUDE_NOTDEFINED_NESTED_SORT
 #endif
 
-#include "sorting/Kokkos_NestedSort.hpp"
+#include "sorting/Kokkos_NestedSortPublicAPI.hpp"
 
 #ifdef KOKKOS_IMPL_PUBLIC_INCLUDE_NOTDEFINED_NESTED_SORT
 #undef KOKKOS_IMPL_PUBLIC_INCLUDE
